refactor features overview table
