+++ admin-frontend/app_singleapp/lib/widgets/features/feature_dashboard_constants.dart
@@ -1,0 +1,3 @@
+final unselectedRowHeight = 60.0;
+final selectedRowHeight = 300.0;
+final headerHeight = 40.0;
