+++ admin-frontend/app_singleapp/lib/widgets/features/feature_value_row_generic.dart
@@ -26,20 +26,31 @@ class FeatureValueNameCell extends StatelessWidget {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: <Widget>[
-          Text('Feature key',
-              overflow: TextOverflow.ellipsis,
-              style: Theme.of(context)
-                  .textTheme
-                  .caption
-                  .copyWith(color: Theme.of(context).buttonColor)),
-          SizedBox(height: 2.0),
-          Text(feature.key,
-              overflow: TextOverflow.ellipsis,
-              style: TextStyle(fontFamily: 'source', fontSize: 12)),
-          FHCopyToClipboard(
-            tooltipMessage: 'Copy key',
-            copyString: feature.key,
+          Column(
+            mainAxisAlignment: MainAxisAlignment.start,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text('Feature key',
+                  overflow: TextOverflow.ellipsis,
+                  style: Theme.of(context)
+                      .textTheme
+                      .caption
+                      .copyWith(color: Theme.of(context).buttonColor)),
+              SizedBox(height: 2.0),
+              Row(
+                children: [
+                  Text(feature.key,
+                      overflow: TextOverflow.ellipsis,
+                      style: TextStyle(fontFamily: 'source', fontSize: 12)),
+                  FHCopyToClipboard(
+                    tooltipMessage: 'Copy key',
+                    copyString: feature.key,
+                  ),
+                ],
+              ),
+            ],
           ),
+
           SizedBox(
             height: 12.0,
           ),
